fix: scroll_offset u32 → u64, same type as evicted_lines
Both offset and evicted_lines are now u64, so subtraction never
wraps due to type mismatch. Only converted to u16 at the ratatui
boundary in conversation.rs (clamped to u16::MAX).

+++ src/ui/render_state.rs
@@ -29,7 +29,7 @@ pub struct RenderState {
     pub menu_items: Vec<(String, String)>,
     pub menu_idx: usize,
     pub menu_title: Option<String>,
-    pub scroll_offset: u32,
+    pub scroll_offset: u64,
     pub follow_tail: bool,
     #[allow(dead_code)]
     pub prev_total_lines: u32,
